+++ spec/github_repo_changes_spec.rb
@@ -19,4 +19,12 @@ describe GithubRepoChanges do
     expect(github_changes.old_ref).to eq('master')
     expect(github_changes.new_ref).to eq('develop')
   end
+
+  it 'Check correct message if single ref - is the latest ref' do
+    github_changes.fetch_refs
+    newset_ref = github_changes.refs.first
+    github_changes.refs_from_message("/get_changes #{newset_ref}")
+    expect(github_changes.link_to_changes)
+      .to match(/Your specified version is the latest version/)
+  end
 end